Router refactoring: Application
