+++ backend/api/app/main.py
@@ -6,7 +6,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from starlette.responses import RedirectResponse
 
-from .routers import fam_router
+from .routers import fam_router, router_application
 
 logConfigFile = os.path.join(
     os.path.dirname(__file__),
@@ -70,5 +70,6 @@ def main():
     return RedirectResponse(url="/docs/")
 
 apiPrefix = '/api/v1'
-app.include_router(fam_router.router, prefix=apiPrefix)
+app.include_router(fam_router.router, prefix=apiPrefix) # TODO: this will be gone after router refactoring.
+app.include_router(router_application.router, prefix=apiPrefix + '/fam_applications', tags=["FAM Applications"])
